fix bug

+++ src/main/java/com/example/mux/statistic/model/dto/GroupStepsDTO.java
@@ -2,8 +2,6 @@ package com.example.mux.statistic.model.dto;
 
 import com.example.mux.group.model.Group;
 import com.example.mux.group.model.dto.GroupDTO;
-import com.example.mux.user.model.User;
-import com.example.mux.user.model.dto.UserDTO;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -20,7 +18,7 @@ public class GroupStepsDTO implements Comparable<GroupStepsDTO>{
         setSteps(steps);
         setGroup(new GroupDTO(group));
         if(group.getNumberOfEmployees() > 0){
-            setStepsPerUser(steps / group.getNumberOfEmployees());
+            setStepsPerUser((float) steps / group.getNumberOfEmployees());
         }else{
             setStepsPerUser(0);
         }
